pyproject.toml
[tool.poetry]
name = "rickchurch"
version = "1.0.0"
description = "API for the Church of Rick"
authors = ["ItsDrike <itsdrikeofficial@gmail.com>", "Scoder12 <34356756+Scoder12@users.noreply.github.com>"]
license = "AGPL"

[tool.poetry.dependencies]
python = "3.8.*"
fastapi = "^0.65.1"
pydispix = {git = "https://github.com/ItsDrike/pydispix", rev = "async"}
python-decouple = "^3.4"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
asyncpg = "^0.23.0"

[tool.poetry.dev-dependencies]
autopep8 = "~=1.5.7"
flake8 = "~=3.9.2"
flake8-bugbear = "~=21.4.3"
flake8-import-order = "~=0.18.1"
flake8-tidy-imports = "~=4.3.0"
pep8-naming = "~=0.11.1"
ipython = "~=7.24.0"
pre-commit = "~=2.13.0"
taskipy = "~=1.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
lint = "pre-commit run --all-files"
precommit = "pre-commit install"
